Remove support for Spring WebFlux
Resolves #92
