+++ build.gradle
@@ -96,8 +96,7 @@ subprojects {
 	task dependencyReport(type: DependencyReportTask)
 }
 
-configure(subprojects - [project(":spring-cloud-starter-open-service-broker-webmvc"),
-						 project(":spring-cloud-starter-open-service-broker-webflux")]) {
+configure(subprojects - [project(":spring-cloud-starter-open-service-broker-webmvc")]) {
 	sourceCompatibility = 1.8
 	targetCompatibility = 1.8
 	[compileJava, compileTestJava]*.options*.encoding = "UTF-8"
