+++ docs/src/docs/asciidoc/getting-started.adoc
@@ -3,11 +3,8 @@
 
 Most service broker applications implement API or web UI endpoints beyond the Open Service Broker API endpoints.
 These additional endpoints might provide information about the application, provide a dashboard UI, or provide controls over application behavior.
-Developers can choose to implement these additional endpoints with https://docs.spring.io/spring/docs/current/spring-framework-reference/web.html[Spring MVC] or https://docs.spring.io/spring/docs/current/spring-framework-reference/web-reactive.html[Spring WebFlux].
-
-The choice of Spring web framework does not affect the behavior of the Open Service Broker API endpoints, but the autoconfiguration of the project depends on which web framework the service broker application uses.
-
-Two Spring Boot starter dependencies are provided, reflecting the choice of Spring web framework.
+Developers may implement these additional endpoints with https://docs.spring.io/spring/docs/current/spring-framework-reference/web.html[Spring MVC].
+Spring WebFlux is currently not supported.
 
 === Maven dependencies
 
@@ -21,28 +18,12 @@ To use Spring Cloud Open Service Broker in a *Spring MVC* application, add the `
         </dependency>
     </dependencies>
 
-To use Spring Cloud Open Service Broker in a *Spring WebFlux* application, add the `webflux` starter:
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-starter-open-service-broker-webflux</artifactId>
-            <version>${version}</version>
-        </dependency>
-    </dependencies>
-
 === Gradle dependencies
 
 To use Spring Cloud Open Service Broker in a *Spring MVC* application, add the `webmvc` starter:
 
     dependencies {
         compile("org.springframework.cloud:spring-cloud-starter-open-service-broker-webmvc:${version}")
-    }
-
-To use Spring Cloud Open Service Broker in a *Spring WebFlux* application, add the `webflux` starter:
-
-    dependencies {
-        compile("org.springframework.cloud:spring-cloud-starter-open-service-broker-webflux:${version}")
     }
 
 === Configuring the service broker
