+++ spring-cloud-open-service-broker-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.springframework.cloud.servicebroker.autoconfigure.web.ServiceBrokerAutoConfiguration,\
-org.springframework.cloud.servicebroker.autoconfigure.web.servlet.ServiceBrokerWebMvcAutoConfiguration,\
-org.springframework.cloud.servicebroker.autoconfigure.web.servlet.ApiVersionWebMvcAutoConfiguration,\
-org.springframework.cloud.servicebroker.autoconfigure.web.reactive.ServiceBrokerWebFluxAutoConfiguration,\
-org.springframework.cloud.servicebroker.autoconfigure.web.reactive.ApiVersionWebFluxAutoConfiguration
+org.springframework.cloud.servicebroker.autoconfigure.web.ServiceBrokerWebMvcAutoConfiguration,\
+org.springframework.cloud.servicebroker.autoconfigure.web.ApiVersionWebMvcAutoConfiguration
