+++ spring-cloud-open-service-broker-autoconfigure/src/test/java/org/springframework/cloud/servicebroker/autoconfigure/web/ApiVersionWebMvcAutoConfigurationTest.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.servicebroker.autoconfigure.web.reactive;
+package org.springframework.cloud.servicebroker.autoconfigure.web;
 
 import org.junit.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-import org.springframework.boot.test.context.runner.ReactiveWebApplicationContextRunner;
-import org.springframework.cloud.servicebroker.autoconfigure.web.TestServiceInstanceService;
+import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
 import org.springframework.cloud.servicebroker.model.BrokerApiVersion;
 import org.springframework.cloud.servicebroker.service.ServiceInstanceService;
 import org.springframework.context.annotation.Bean;
@@ -32,7 +31,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.cloud.servicebroker.model.BrokerApiVersion.API_VERSION_ANY;
 import static org.springframework.cloud.servicebroker.model.BrokerApiVersion.API_VERSION_CURRENT;
 
-public class ApiVersionWebFluxAutoConfigurationTest {
+public class ApiVersionWebMvcAutoConfigurationTest {
 
 	@Test
 	public void apiVersionBeansAreNotCreatedWithNonWebConfiguration() {
@@ -40,7 +39,8 @@ public class ApiVersionWebFluxAutoConfigurationTest {
 				.withUserConfiguration(ServicesConfiguration.class)
 				.run((context) -> {
 					assertThat(context).doesNotHaveBean(BrokerApiVersion.class);
-					assertThat(context).doesNotHaveBean(ApiVersionWebFilter.class);
+					assertThat(context).doesNotHaveBean(ApiVersionInterceptor.class);
+					assertThat(context).doesNotHaveBean(ApiVersionWebMvcConfigurerAdapter.class);
 				});
 	}
 
@@ -49,18 +49,8 @@ public class ApiVersionWebFluxAutoConfigurationTest {
 		webApplicationContextRunner()
 				.run((context) -> {
 					assertThat(context).doesNotHaveBean(BrokerApiVersion.class);
-					assertThat(context).doesNotHaveBean(ApiVersionWebFilter.class);
-				});
-	}
-
-	@Test
-	public void apiVersionCheckIsDisabled() {
-		webApplicationContextRunner()
-				.withUserConfiguration(ServicesConfiguration.class)
-				.withPropertyValues("spring.cloud.openservicebroker.apiVersionCheckEnabled=false")
-				.run((context) -> {
-					assertThat(context).doesNotHaveBean(BrokerApiVersion.class);
-					assertThat(context).doesNotHaveBean(ApiVersionWebFilter.class);
+					assertThat(context).doesNotHaveBean(ApiVersionInterceptor.class);
+					assertThat(context).doesNotHaveBean(ApiVersionWebMvcConfigurerAdapter.class);
 				});
 	}
 
@@ -71,18 +61,30 @@ public class ApiVersionWebFluxAutoConfigurationTest {
 				.run((context) -> {
 					assertThat(context).getBean(BrokerApiVersion.class)
 							.hasFieldOrPropertyWithValue("apiVersion", API_VERSION_ANY);
-					assertThat(context).hasSingleBean(ApiVersionWebFilter.class);
+					assertThat(context).hasSingleBean(ApiVersionInterceptor.class);
+					assertThat(context).hasSingleBean(ApiVersionWebMvcConfigurerAdapter.class);
 				});
 	}
 
 	@Test
-	public void apiVersionBeansAreCreatedWithCustomVersion() {
+	public void apiVersionCheckIsDisabled() {
+		webApplicationContextRunner()
+				.withUserConfiguration(ServicesConfiguration.class)
+				.withPropertyValues("spring.cloud.openservicebroker.apiVersionCheckEnabled=false")
+				.run((context) -> {
+					assertThat(context).doesNotHaveBean(BrokerApiVersion.class);
+				});
+	}
+
+	@Test
+	public void apiVersionBeansAreCreatedFromCustomVersionBean() {
 		webApplicationContextRunner()
 				.withUserConfiguration(ServicesConfiguration.class, CustomBrokerApiVersionConfigurationFromBean.class)
 				.run((context) -> {
 					assertThat(context).getBean(BrokerApiVersion.class)
 							.hasFieldOrPropertyWithValue("apiVersion", API_VERSION_CURRENT);
-					assertThat(context).hasSingleBean(ApiVersionWebFilter.class);
+					assertThat(context).hasSingleBean(ApiVersionInterceptor.class);
+					assertThat(context).hasSingleBean(ApiVersionWebMvcConfigurerAdapter.class);
 				});
 	}
 
@@ -93,7 +95,8 @@ public class ApiVersionWebFluxAutoConfigurationTest {
 				.run((context) -> {
 					assertThat(context).getBean(BrokerApiVersion.class)
 							.hasFieldOrPropertyWithValue("apiVersion", "42.321");
-					assertThat(context).hasSingleBean(ApiVersionWebFilter.class);
+					assertThat(context).hasSingleBean(ApiVersionInterceptor.class);
+					assertThat(context).hasSingleBean(ApiVersionWebMvcConfigurerAdapter.class);
 				});
 	}
 
@@ -104,18 +107,19 @@ public class ApiVersionWebFluxAutoConfigurationTest {
 				.run((context) -> {
 					assertThat(context).getBean(BrokerApiVersion.class)
 							.hasFieldOrPropertyWithValue("apiVersion", "99.999");
-					assertThat(context).hasSingleBean(ApiVersionWebFilter.class);
+					assertThat(context).hasSingleBean(ApiVersionInterceptor.class);
+					assertThat(context).hasSingleBean(ApiVersionWebMvcConfigurerAdapter.class);
 				});
 	}
 
-	private ReactiveWebApplicationContextRunner webApplicationContextRunner() {
-		return new ReactiveWebApplicationContextRunner()
-				.withConfiguration(AutoConfigurations.of(ApiVersionWebFluxAutoConfiguration.class));
+	private WebApplicationContextRunner webApplicationContextRunner() {
+		return new WebApplicationContextRunner()
+				.withConfiguration(AutoConfigurations.of(ApiVersionWebMvcAutoConfiguration.class));
 	}
 
 	private ApplicationContextRunner nonWebApplicationContextRunner() {
 		return new ApplicationContextRunner()
-				.withConfiguration(AutoConfigurations.of(ApiVersionWebFluxAutoConfiguration.class));
+				.withConfiguration(AutoConfigurations.of(ApiVersionWebMvcAutoConfiguration.class));
 	}
 
 	@Configuration
