--- spring-cloud-open-service-broker-autoconfigure/src/test/resources/contracts/catalog/reactive/retrieveCatalog.groovy
@@ -1,33 +1,0 @@
-/*
- * Copyright 2002-2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package catalog.reactive
-
-org.springframework.cloud.contract.spec.Contract.make {
-	request {
-		method 'GET'
-		url '/v2/catalog'
-		headers {
-			contentType('application/json')
-		}
-	}
-	response {
-		status 200
-		body(file("retrieve_catalog_response.json"))
-		headers {
-			contentType('application/json')
-		}
-	}
-}
