+++ spring-cloud-open-service-broker-core/build.gradle
@@ -25,7 +25,6 @@ dependencyManagement {
 dependencies {
 	compile("org.springframework:spring-context")
 	optional("org.springframework:spring-web")
-	optional("org.springframework:spring-webflux")
 
 	compile("com.fasterxml.jackson.core:jackson-databind:2.9.5")
 	compile("org.hibernate.validator:hibernate-validator:6.0.9.Final")
